+++ roipoly/roipoly.py
@@ -79,8 +79,23 @@ class RoiPoly:
         else:
             plt.show(block=True)
 
-    def get_mask(self, current_image):
-        ny, nx = np.shape(current_image)[0:2]
+
+    def get_mask(self, image):
+        """Get binary mask of the ROI polygon.
+
+        Parameters
+        ----------
+        image: numpy array (2D)
+            Image that the mask should be based on. Only used for determining
+            the shape of the binary mask (which is made equal to the shape of
+            the image)
+
+        Returns
+        -------
+        numpy array (2D)
+
+        """
+        ny, nx = np.shape(image)
         poly_verts = ([(self.x[0], self.y[0])]
                       + list(zip(reversed(self.x), reversed(self.y))))
         # Create vertex coordinates for each grid cell...
@@ -90,9 +105,9 @@ class RoiPoly:
         points = np.vstack((x, y)).T
 
         roi_path = MplPath(poly_verts)
-        grid = roi_path.contains_points(points).reshape((ny, nx))
-        grid=grid if current_image.ndim!=3 else np.dstack(tuple([grid for i in range(current_image.shape[2])]))
-        return grid
+
+        mask = roi_path.contains_points(points).reshape((ny, nx))
+        return mask
 
     def display_roi(self, **linekwargs):
         line = plt.Line2D(self.x + [self.x[0]], self.y + [self.y[0]],
@@ -102,14 +117,39 @@ class RoiPoly:
         plt.draw()
         self.show_figure()
 
-    def get_mean_and_std(self, current_image):
-        mask = self.get_mask(current_image)
-        mean = np.mean(np.extract(mask, current_image))
-        std = np.std(np.extract(mask, current_image))
+    def get_mean_and_std(self, image):
+        """Get statistics about pixel values of an image inside the ROI.
+
+        Parameters
+        ----------
+        image: numpy array (2D)
+            Image on which the statistics should be calculated
+
+        Returns
+        -------
+        list of float:
+            mean and standard deviation of the pixel values inside the ROI
+
+        """
+        mask = self.get_mask(image)
+        mean = np.mean(np.extract(mask, image))
+        std = np.std(np.extract(mask, image))
         return mean, std
 
-    def display_mean(self, current_image, **textkwargs):
-        mean, std = self.get_mean_and_std(current_image)
+    def display_mean(self, image, **textkwargs):
+        """Display statistics about pixel values of an image inside the ROI.
+
+        Parameters
+        ----------
+        image: numpy array (2D)
+            Image on which the statistics should be calculated
+
+        Returns
+        -------
+        None
+
+        """
+        mean, std = self.get_mean_and_std(image)
         string = "%.3f +- %.3f" % (mean, std)
         plt.text(self.x[0], self.y[0],
                  string, color=self.color,
@@ -131,7 +171,7 @@ class RoiPoly:
         if event.inaxes == self.ax:
             x, y = event.xdata, event.ydata
             ax = event.inaxes
-            if event.button == 1 and event.dblclick is False:
+            if event.button == 1 and not event.dblclick:
                 logger.debug("Received single left mouse button click")
                 if self.line is None:  # If there is no line, create a line
                     self.line = plt.Line2D([x, x], [y, y],
@@ -159,8 +199,8 @@ class RoiPoly:
                     event.inaxes.add_line(self.line)
                     self.fig.canvas.draw()
 
-            elif (((event.button == 1 and event.dblclick is True) or
-                   (event.button == 3 and event.dblclick is False)) and
+            elif (((event.button == 1 and event.dblclick) or
+                   (event.button == 3 and not event.dblclick)) and
                   self.line is not None):
                 # Close the loop and disconnect
                 logger.debug("Received single right mouse button click or "